factor code into Runner - enable testing
